Add constant register awareness to compiler
It has been decided for the Leros ABI to include a set of constant registers. These may be used by the custom inserters as well as be selected during instruction selection if they contain a required constant.

The registers are currently initialized through crt0.leros.c. This may also be done by an OS, removing the need for all programs to set the constant register state.

Reduces CoreMark .text section size by ~10%